Rebalance nitro usage and kart classes
- The bonus nitro duration (after consumption is stopped) is reduced from 1.0 to 0.8
- Big nitro cans give 2.5 instead of 3.0
- Light karts get a max speed boost of 6 instead of 4.5
- Medium karts get a max speed boost of 5.5 instead of 5
- Light and medium karts nitro consumption are slightly increased to balance the speed bonus
- The minimum nitro burst is changed from 8/15th of a second to 0.5s

It was too common for light and medium karts to collect nitro faster than they could use it, this issue should be alleviated.

Furthermore, switch the min_nitro_consumption to an int16 to allow values greater than 64 (required with the physics FPS change).

+++ src/karts/kart.hpp
@@ -210,7 +210,7 @@ protected:
     float        m_current_lean;
 
     /** To prevent using nitro in too short bursts */
-    int8_t        m_min_nitro_ticks;
+    int16_t        m_min_nitro_ticks;
 
     /** True if fire button was pushed and not released */
     bool         m_fire_clicked;
